@@ -27,11 +27,11 @@ project_urls =
 [options]
 packages = find:
 install_requires =
-    matplotlib<3.8
+    matplotlib
     napari
     numpy
     tinycss2
-python_requires = >=3.8
+python_requires = >=3.9
 include_package_data = True
 package_dir =
     =src
@@ -49,6 +49,7 @@ napari.manifest =
 docs =
     napari[all]==0.4.17
     numpydoc
+    pydantic<2
     pydata-sphinx-theme
     qtgallery
     sphinx
